requirements.txt
beautifulsoup4
pyppeteer
fastapi[all]
uvicorn
python-dotenv
groq
boto3
datetime
requests
Pillow
sentencepiece
transformers
asyncpg